--- src/components/content/content_alteredchromatic.tsx
@@ -1,11 +1,0 @@
-function Content() {
-  return (
-    <>
-      <h2>{'AlteredChromatic'}</h2>
-      <h3>{'Premise'}</h3>
-      <p>{'What if music was simpler, and a piano had a black key for every white key?'}</p>
-    </>
-  );
-}
-
-export default Content;
