--- src/components/content/content_browserpet.tsx
@@ -1,11 +1,0 @@
-function Content() {
-  return (
-    <>
-      <h2>{'BrowserPet'}</h2>
-      <h3>{'Premise'}</h3>
-      <p>{'Keep some dumb animals alive in a chrome extension'}</p>
-    </>
-  );
-}
-
-export default Content;
