--- src/components/content/content_dropship.tsx
@@ -1,23 +1,0 @@
-function Content() {
-  return (
-    <>
-      <h2>{'DROPSHIP'}</h2>
-      <h3>{'Premise'}</h3>
-      <p>{'A roguelike shooter build in GameMaker studio'}</p>
-      <p>
-        {
-          'Spin the reels of fortune in a world of slot machines, where excitement and anticipation reign supreme. The clinking of coins, the flashing lights, and the thrilling soundtracks create an immersive gaming experience that keeps players coming back for more.'
-        }
-      </p>
-      <ul>
-        <li>
-          {'In this realm of chance, every pull of the lever or press of the button represents a shot at big winnings.'}
-        </li>
-        <li>{`Whether you're chasing cherries, lucky sevens, or bonus symbols, the spinning reels hold the promise of jackpots and thrilling payouts.`}</li>
-        <li>{'With each spin, the heart races, and dreams of hitting the ultimate jackpot come to life.'}</li>
-      </ul>
-    </>
-  );
-}
-
-export default Content;
