--- src/components/content/content_fretref.tsx
@@ -1,11 +1,0 @@
-function Content() {
-  return (
-    <>
-      <h2>{'FretRef'}</h2>
-      <h3>{'Premise'}</h3>
-      <p>{'An app that helps you play tasty licks without all the hard work'}</p>
-    </>
-  );
-}
-
-export default Content;
