--- src/components/content/content_raccoontrapper.tsx
@@ -1,12 +1,0 @@
-function Content() {
-  return (
-    <>
-      <h2>{'Raccoon Trapper'}</h2>
-      <h3>{'Premise'}</h3>
-      <p>{'A twist on the classic game, built in PhaserJS'}</p>
-      <p>{'feed some cute critters, but be careful, they get hangry!'}</p>
-    </>
-  );
-}
-
-export default Content;
